--- app/src/main/java/pcs/ub/edu/ar/clinicavirtual/data/AppointmentItemListener.java
@@ -1,4 +1,0 @@
-package pcs.ub.edu.ar.clinicavirtual.data;
-
-class AppointmentItemListener {
-}
