+++ app/src/main/java/pcs/ub/edu/ar/clinicavirtual/data/appointment/AppointmentItemListener.java
@@ -1,0 +1,4 @@
+package pcs.ub.edu.ar.clinicavirtual.data.appointment;
+
+class AppointmentItemListener {
+}
