+++ app/src/main/java/pcs/ub/edu/ar/clinicavirtual/data/appointment/PagingAppointmentCriteria.java
@@ -1,7 +1,10 @@
-package pcs.ub.edu.ar.clinicavirtual.data;
+package pcs.ub.edu.ar.clinicavirtual.data.appointment;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import pcs.ub.edu.ar.clinicavirtual.data.Appointment;
+import pcs.ub.edu.ar.clinicavirtual.data.appointment.interfaces.IAppointmentCriteria;
 
 public class PagingAppointmentCriteria implements IAppointmentCriteria {
