+++ app/src/main/java/pcs/ub/edu/ar/clinicavirtual/data/appointment/datasource/MemoryAppointmentsDataSource.java
@@ -1,4 +1,4 @@
-package pcs.ub.edu.ar.clinicavirtual.data;
+package pcs.ub.edu.ar.clinicavirtual.data.appointment.datasource;
 
 import com.google.common.collect.Lists;
 
@@ -6,6 +6,10 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
+
+import pcs.ub.edu.ar.clinicavirtual.data.Appointment;
+import pcs.ub.edu.ar.clinicavirtual.data.appointment.interfaces.IAppointmentCriteria;
+import pcs.ub.edu.ar.clinicavirtual.data.appointment.interfaces.IMemoryAppointmentsDataSource;
 
 public class MemoryAppointmentsDataSource implements IMemoryAppointmentsDataSource {
