+++ app/src/main/java/pcs/ub/edu/ar/clinicavirtual/data/appointment/interfaces/IAppointmentsRepository.java
@@ -1,6 +1,8 @@
-package pcs.ub.edu.ar.clinicavirtual.data;
+package pcs.ub.edu.ar.clinicavirtual.data.appointment.interfaces;
 
 import java.util.List;
+
+import pcs.ub.edu.ar.clinicavirtual.data.Appointment;
 
 public interface IAppointmentsRepository {
     interface GetAppointmentsCallback{
