+++ app/src/main/java/pcs/ub/edu/ar/clinicavirtual/handler/GetPatientAppointmentsHandler.java
@@ -1,19 +1,12 @@
 package pcs.ub.edu.ar.clinicavirtual.handler;
 
-import android.widget.Toast;
-
-import org.json.JSONArray;
 import org.json.JSONException;
-import org.json.JSONObject;
 
 import pcs.ub.edu.ar.clinicavirtual.activitys.base.BaseActivity;
 import pcs.ub.edu.ar.clinicavirtual.activitys.general.MyTurnsActivity;
 import pcs.ub.edu.ar.clinicavirtual.connection.facade.pattern.connection.requests.user.ServerRequestUserGetPatientAppointments;
-import pcs.ub.edu.ar.clinicavirtual.data.CloudAppointmentsDatasource;
 import pcs.ub.edu.ar.clinicavirtual.interfaces.handler.IServerResponseHandler;
 import pcs.ub.edu.ar.clinicavirtual.interfaces.facade.pattern.connection.IServerRequest;
-
-import static pcs.ub.edu.ar.clinicavirtual.data.CloudAppointmentsDatasource.*;
 
 public class GetPatientAppointmentsHandler implements IServerResponseHandler {
